refactor(ai): 重构CopilotWidget组件及聊天消息相关代码
- 优化聊天消息组件为UserMessage和AssistantMessage两个memo组件
- 优化消息生成逻辑，新增生成消息ID的辅助函数
- 改进聊天消息渲染，支持流式消息和带引用来源展示
- 简化消息UI类型判断，统一处理解释型和非解释型内容
- 新增建议动作按钮列表组件SuggestedActionList
- 重构ChatMessage组件为函数组件，提升渲染性能和可读性
- 修正索引服务调用，传递完整用户对象参数
- 在身份认证页面新增OTP登录模式及邮箱验证接口调用
- 清理身份认证接口代码，删除多余日志打印
- 移除部分无用条件判断，修正接口返回逻辑
- 修改同步到云端时文章过滤条件，移除deletedAt约束

+++ prisma/migrations/20250116_fix_book_schema/migration.sql
@@ -170,7 +170,7 @@ SET completed_chapters = (
 UPDATE collections c
 SET reading_progress = CASE
     WHEN c.total_chapters > 0 THEN
-        ROUND((c.completed_chapters::FLOAT / c.total_chapters::FLOAT) * 100, 2)
+        ROUND((c.completed_chapters::FLOAT / c.total_chapters::FLOAT) * 100::NUMERIC, 2)
     ELSE 0
 END
 WHERE c.total_chapters > 0;
@@ -207,7 +207,7 @@ BEGIN
         -- 更新进度百分比
         UPDATE collections
         SET reading_progress = ROUND(
-            (completed_chapters::FLOAT / GREATEST(total_chapters, 1)::FLOAT) * 100, 2
+            (completed_chapters::FLOAT / GREATEST(total_chapters, 1)::FLOAT) * 100::NUMERIC, 2
         )
         WHERE id = NEW.collection_id;
 
@@ -271,7 +271,7 @@ SELECT
     c.reading_progress,
     COUNT(a.id) FILTER (WHERE a.progress > 0 AND a.progress < 99) AS in_progress_chapters,
     COUNT(a.id) FILTER (WHERE a.progress >= 99) AS completed_count,
-    ROUND(AVG(a.progress), 2) AS avg_article_progress
+    ROUND(AVG(a.progress)::NUMERIC, 2) AS avg_article_progress
 FROM collections c
 LEFT JOIN articles a ON a.collection_id = c.id AND a.deleted_at IS NULL
 GROUP BY c.id, c.title, c.type, c.total_chapters, c.completed_chapters, c.reading_progress;
